refactor: isolate call recovery monitor
