+++ TelnyxRTC/Telnyx/Services/CallRecovery/SignalingHealthMonitor.swift
@@ -1,0 +1,155 @@
+import Foundation
+import WebRTC
+
+/// Decides whether a failed active call restarts ICE or uses reconnect/reattach.
+///
+/// The monitor owns recovery state and timers, while `Call` owns WebRTC
+/// renegotiation and `TxClient` owns reconnect/reattach mechanics.
+internal final class SignalingHealthMonitor {
+    private enum RecoveryMode: Equatable {
+        case idle
+        case iceRestarting
+        case reattaching
+    }
+
+    private let iceRestartTimeout: TimeInterval
+    private let isSignalingAvailable: () -> Bool
+    private let startIceRestart: (Call) -> Void
+    private let requestReattach: () -> Void
+
+    private weak var recoveringCall: Call?
+    private var recoveryMode: RecoveryMode = .idle
+    private var iceRestartTimeoutWorkItem: DispatchWorkItem?
+
+    init(
+        iceRestartTimeout: TimeInterval = 15,
+        isSignalingAvailable: @escaping () -> Bool,
+        startIceRestart: @escaping (Call) -> Void,
+        requestReattach: @escaping () -> Void
+    ) {
+        self.iceRestartTimeout = iceRestartTimeout
+        self.isSignalingAvailable = isSignalingAvailable
+        self.startIceRestart = startIceRestart
+        self.requestReattach = requestReattach
+    }
+
+    func networkPathDidChange() {
+        executeOnMain { [weak self] in
+            guard let self = self, self.recoveryMode != .reattaching else { return }
+            Logger.log.i(message: "[CALL-RECOVERY] Network path changed; using required reconnect/reattach flow")
+            self.cancelIceRestartTimeout()
+            self.recoveringCall = nil
+            self.recoveryMode = .reattaching
+        }
+    }
+
+    func iceConnectionStateDidChange(_ call: Call, state: RTCIceConnectionState) {
+        guard state == .failed else { return }
+        requestRecovery(for: call, trigger: "ice_failed")
+    }
+
+    func peerConnectionStateDidChange(_ call: Call, state: RTCPeerConnectionState) {
+        guard state == .failed else { return }
+        requestRecovery(for: call, trigger: "peer_connection_failed")
+    }
+
+    func iceRestartDidComplete(for call: Call) {
+        executeOnMain { [weak self] in
+            guard let self = self, self.recoveryMode == .iceRestarting, self.recoveringCall === call else { return }
+            Logger.log.i(message: "[CALL-RECOVERY] ICE restart answer applied")
+            self.finishRecovery()
+        }
+    }
+
+    func iceRestartRequestDidFail(for call: Call, error: Error) {
+        executeOnMain { [weak self] in
+            guard let self = self, self.recoveryMode == .iceRestarting, self.recoveringCall === call else { return }
+            Logger.log.e(message: "[CALL-RECOVERY] ICE restart failed: \(error.localizedDescription)")
+            self.beginReattach()
+        }
+    }
+
+    func callStateDidChange(_ call: Call) {
+        executeOnMain { [weak self] in
+            guard let self = self else { return }
+
+            switch call.callState {
+            case .ACTIVE:
+                if self.recoveryMode == .reattaching {
+                    Logger.log.i(message: "[CALL-RECOVERY] Reattached call is active")
+                    self.finishRecovery()
+                }
+            case .DONE, .DROPPED:
+                if self.recoveringCall === call {
+                    self.finishRecovery()
+                }
+            default:
+                break
+            }
+        }
+    }
+
+    private func requestRecovery(for call: Call, trigger: String) {
+        executeOnMain { [weak self] in
+            guard let self = self else { return }
+            guard call.callState == .ACTIVE else {
+                Logger.log.i(message: "[CALL-RECOVERY] Ignoring \(trigger); call is \(call.callState.value)")
+                return
+            }
+            guard self.recoveryMode == .idle else {
+                Logger.log.i(message: "[CALL-RECOVERY] Ignoring \(trigger); recovery is already in progress")
+                return
+            }
+
+            self.recoveringCall = call
+            if self.isSignalingAvailable() {
+                Logger.log.w(message: "[CALL-RECOVERY] \(trigger) with signaling available; starting ICE restart")
+                self.recoveryMode = .iceRestarting
+                self.startIceRestartTimeout(for: call)
+                self.startIceRestart(call)
+            } else {
+                Logger.log.w(message: "[CALL-RECOVERY] \(trigger) with signaling unavailable; starting reconnect/reattach")
+                self.beginReattach()
+            }
+        }
+    }
+
+    private func startIceRestartTimeout(for call: Call) {
+        cancelIceRestartTimeout()
+        let workItem = DispatchWorkItem { [weak self, weak call] in
+            guard let self = self,
+                  let call = call,
+                  self.recoveryMode == .iceRestarting,
+                  self.recoveringCall === call else { return }
+            Logger.log.e(message: "[CALL-RECOVERY] ICE restart timed out after \(self.iceRestartTimeout)s")
+            self.beginReattach()
+        }
+        iceRestartTimeoutWorkItem = workItem
+        DispatchQueue.main.asyncAfter(deadline: .now() + iceRestartTimeout, execute: workItem)
+    }
+
+    private func beginReattach() {
+        cancelIceRestartTimeout()
+        recoveryMode = .reattaching
+        requestReattach()
+    }
+
+    private func finishRecovery() {
+        cancelIceRestartTimeout()
+        recoveringCall = nil
+        recoveryMode = .idle
+    }
+
+    private func cancelIceRestartTimeout() {
+        iceRestartTimeoutWorkItem?.cancel()
+        iceRestartTimeoutWorkItem = nil
+    }
+
+    private func executeOnMain(_ block: @escaping () -> Void) {
+        if Thread.isMainThread {
+            block()
+        } else {
+            DispatchQueue.main.async(execute: block)
+        }
+    }
+}
